cpass integrate round 5: 8 groups, 0 fields merged, 42 fixes

+++ crates/mp/engine/botlib/src/be_aas_reach_fns.rs
@@ -819,7 +819,7 @@ pub fn AAS_SetupReachabilityHeap(bot: &mut BotLib) {
     unsafe {
         bot.reachabilityheap = GetClearedMemory(
             bot,
-            AAS_MAX_REACHABILITYSIZE as usize * core::mem::size_of::<aas_lreachability_t>(),
+            AAS_MAX_REACHABILITYSIZE as u64 * core::mem::size_of::<aas_lreachability_t>() as u64,
         ) as *mut aas_lreachability_t;
         let mut i = 0;
         while i < AAS_MAX_REACHABILITYSIZE - 1 {
@@ -1319,7 +1319,7 @@ pub fn AAS_StoreReachability(bot: &mut BotLib) {
         }
         bot.aasworld.reachability = GetClearedMemory(
             bot,
-            (bot.numlreachabilities + 10) as usize * core::mem::size_of::<aas_reachability_t>(),
+            (bot.numlreachabilities + 10) as u64 * core::mem::size_of::<aas_reachability_t>() as u64,
         ) as *mut aas_reachability_t;
         bot.aasworld.reachabilitysize = 1;
         let mut i = 0;
@@ -3916,7 +3916,7 @@ pub fn AAS_InitReachability(bot: &mut BotLib) {
         //allocate area reachability link array
         bot.areareachability = GetClearedMemory(
             bot,
-            bot.aasworld.numareas as usize * core::mem::size_of::<*mut aas_lreachability_t>(),
+            bot.aasworld.numareas as u64 * core::mem::size_of::<*mut aas_lreachability_t>() as u64,
         ) as *mut *mut aas_lreachability_t;
         //
         AAS_SetWeaponJumpAreaFlags(bot);
